requirements.txt
google-generativeai 
python-dotenv 
uvicorn 
fastapi
websockets
twilio
aiohttp
